+++ services/reviews/deleteReview.js
@@ -5,6 +5,10 @@ import { v4 as uuid } from 'uuid'
 const deleteReview = async (id) => {
   const prisma = new PrismaClient()
   try{
+    const review = await prisma.review.findUnique({ where: { id: id } });
+    if (!review) {
+      throw new Error("User not found");
+    }
     const deletedReview = await prisma.review.delete({
       where: {
         id: id,
