+++ services/users/deleteUser.js
@@ -4,6 +4,10 @@ import { v4 as uuid } from 'uuid'
 async function deleteUser(userId) {
   const prisma = new PrismaClient()
   try{
+    const user = await prisma.user.findUnique({ where: { id: userId } });
+    if (!user) {
+      throw new Error("User not found");
+    }
     const deletedUser = await prisma.user.delete({
     where: {
       id: userId,
